Added logic to add/remove tickets and delete reservations along with e2e testing

+++ backend/src/event/events.service.ts
@@ -3,14 +3,13 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Events } from './events.entity';
 import { EventsDto } from './events.dto';
-import { Price } from 'price/price.entity';
 
 @Injectable()
 export class EventsService {
   constructor(
     @InjectRepository(Events)
-    private readonly eventRepository: Repository<Events>,
-  ) {}
+    private readonly eventRepository: Repository<Events>
+  ) { }
 
   async createEvent(event: EventsDto) {
     const event_to_save = {
@@ -22,6 +21,10 @@ export class EventsService {
   }
 
   async deleteEvent(id: number) {
+    const event = await this.eventRepository.findOne(1);
+    if (!event) {
+      return;
+    }
     await this.eventRepository.delete(id);
     return id;
   }
@@ -30,6 +33,7 @@ export class EventsService {
     const dbEvent = await this.eventRepository.findOne(id);
     if (dbEvent) {
       const event_to_update = {
+        ...dbEvent,
         ...event,
         age_group_limits: event.age_group_limits.join(','),
       };
